+++ src/main/java/com/ing/data/cassandra/jdbc/CassandraConnection.java
@@ -18,6 +18,8 @@ package com.ing.data.cassandra.jdbc;
 import com.datastax.oss.driver.api.core.ConsistencyLevel;
 import com.datastax.oss.driver.api.core.DefaultConsistencyLevel;
 import com.datastax.oss.driver.api.core.Version;
+import com.datastax.oss.driver.api.core.config.DefaultDriverOption;
+import com.datastax.oss.driver.api.core.config.DriverExecutionProfile;
 import com.datastax.oss.driver.api.core.metadata.Metadata;
 import com.datastax.oss.driver.api.core.session.Session;
 import com.google.common.collect.Maps;
@@ -34,13 +36,16 @@ import java.sql.SQLNonTransientConnectionException;
 import java.sql.SQLTimeoutException;
 import java.sql.SQLWarning;
 import java.sql.Statement;
+import java.util.ArrayList;
 import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
 import java.util.Properties;
 import java.util.Set;
 import java.util.TreeSet;
 import java.util.concurrent.ConcurrentMap;
 import java.util.concurrent.ConcurrentSkipListSet;
+import java.util.stream.Collectors;
 
 import static com.ing.data.cassandra.jdbc.CassandraResultSet.DEFAULT_CONCURRENCY;
 import static com.ing.data.cassandra.jdbc.CassandraResultSet.DEFAULT_HOLDABILITY;
@@ -125,6 +130,8 @@ public class CassandraConnection extends AbstractConnection implements Connectio
     public CassandraConnection(final SessionHolder sessionHolder) throws SQLException {
         this.sessionHolder = sessionHolder;
         final Properties sessionProperties = sessionHolder.properties;
+        final DriverExecutionProfile defaultConfigProfile =
+            sessionHolder.session.getContext().getConfig().getDefaultProfile();
         this.debugMode = Boolean.TRUE.toString().equals(sessionProperties.getProperty(TAG_DEBUG, StringUtils.EMPTY));
         this.hostListPrimary = new TreeSet<>();
         this.hostListBackup = new TreeSet<>();
@@ -132,18 +139,24 @@ public class CassandraConnection extends AbstractConnection implements Connectio
         this.clientInfo = new Properties();
         this.url = PROTOCOL.concat(createSubName(sessionProperties));
         this.currentKeyspace = sessionProperties.getProperty(TAG_DATABASE_NAME);
-        this.username = sessionProperties.getProperty(TAG_USER, StringUtils.EMPTY);
+        this.username = sessionProperties.getProperty(TAG_USER,
+            defaultConfigProfile.getString(DefaultDriverOption.AUTH_PROVIDER_USER_NAME, StringUtils.EMPTY));
         final String cqlVersion = sessionProperties.getProperty(TAG_CQL_VERSION, DEFAULT_CQL_VERSION);
         this.connectionProperties.setProperty(TAG_ACTIVE_CQL_VERSION, cqlVersion);
         this.defaultConsistencyLevel = DefaultConsistencyLevel.valueOf(
-            sessionProperties.getProperty(TAG_CONSISTENCY_LEVEL, ConsistencyLevel.ONE.name()));
+            sessionProperties.getProperty(TAG_CONSISTENCY_LEVEL,
+                defaultConfigProfile.getString(DefaultDriverOption.REQUEST_CONSISTENCY,
+                    ConsistencyLevel.LOCAL_ONE.name())));
         this.cSession = sessionHolder.session;
         this.metadata = cSession.getMetadata();
 
-        LOG.info(String.format("Connected to cluster: %s, with session: %s", getCatalog(), cSession.getName()));
+        final List<SessionHolder> l = new ArrayList<>();
+        l.stream().map(s -> s.session).collect(Collectors.toList());
+
+        LOG.info("Connected to cluster: {}, with session: {}", getCatalog(), cSession.getName());
         metadata.getNodes().forEach(
-            (uuid, node) -> LOG.info(String.format("Datacenter: %s; Host: %s; Rack: %s", node.getDatacenter(),
-                node.getEndPoint().resolve(), node.getRack()))
+            (uuid, node) -> LOG.info("Datacenter: {}; Host: {}; Rack: {}", node.getDatacenter(),
+                node.getEndPoint().resolve(), node.getRack())
         );
 
         // TODO this is shared among all Connections, what if they belong to different clusters?
@@ -153,8 +166,7 @@ public class CassandraConnection extends AbstractConnection implements Connectio
                 CassandraConnection.dbMajorVersion = cassandraVersion.getMajor();
                 CassandraConnection.dbMinorVersion = cassandraVersion.getMinor();
                 CassandraConnection.dbPatchVersion = cassandraVersion.getPatch();
-                LOG.info(String.format("Node: %s runs Cassandra v.%s", entry.getValue().getEndPoint().resolve(),
-                    cassandraVersion.toString()));
+                LOG.info("Node: {} runs Cassandra v.{}", entry.getValue().getEndPoint().resolve(), cassandraVersion);
             }
         });
     }