+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 ### Added
 - Add support for connecting to Cassandra DBaaS cluster with secure connect bundle. 
   See the feature request [#1](https://github.com/ing-bank/cassandra-jdbc-wrapper/discussions/1).
+- Add query parameter `configfile` to use a configuration file instead of the settings defined in the JDBC URL.
 ### Changed
 - Update DataStax Java Driver for Apache Cassandra(R) to version 4.13.0.
 - Update Apache Commons Lang to version 3.12.0.